Rename graphql middlewares + move healthz
